+++ changelog.md
@@ -40,6 +40,11 @@ Initial
  + Fixed a bug where the updater filepaths wouldn't correctly resolve if the console was not in the nrelay directory.
 
 ## `3.3.0`
-## Changes:
+### Changes:
  + Added `Objects.xml` and `GrounTypes.xml` to updater.
  + Added `resources/` to `.gitignore`. Manually change your version to force an update which will download the latest xml resources.
+
+## `3.4.0`
+### Changes:
+ + Added `ReconnectPacket` and `UsePortalPacket`.
+ + Added reconnect packet hook on `Client`.